Consolidate path logic
