+++ .vscode/settings.json
@@ -5,5 +5,6 @@
     "**/bower_components": true,
     "webapp/static": true,
   },
-  "debug.node.autoAttach": "on"
+  "debug.node.autoAttach": "on",
+  "go.formatTool": "goimports"
 }
