+++ webapp/components/test/path-part.html
@@ -5,7 +5,7 @@
   <script src="../../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
   <script src="../../node_modules/wct-browser-legacy/browser.js"></script>
 
-  <script type="module" src="../path-part.js"></script>
+  <script type="module" src="../path.js"></script>
 </head>
 <body>
   <test-fixture id="path-part-dir-prefixed-fixture">
@@ -32,7 +32,7 @@
     </template>
   </test-fixture>
   <script type="module">
-import { PathPart } from '../path-part.js';
+import { PathPart } from '../path.js';
 import { PolymerElement } from '../../node_modules/@polymer/polymer/polymer-element.js';
 
 suite('PathPart', () => {
