+++ webapp/views/wpt-app.js
@@ -1,20 +1,21 @@
+import { PathInfo } from '../components/path.js';
 import '../components/test-runs-query-builder.js';
 import { TestRunsUIQuery } from '../components/test-runs-query.js';
 import '../components/test-search.js';
+import '../components/wpt-flags.js';
 import { WPTFlags } from '../components/wpt-flags.js';
 import '../components/wpt-header.js';
 import '../components/wpt-permalinks.js';
-import '../components/wpt-flags.js';
 import '../node_modules/@polymer/app-route/app-location.js';
 import '../node_modules/@polymer/app-route/app-route.js';
 import '../node_modules/@polymer/iron-pages/iron-pages.js';
 import '../node_modules/@polymer/polymer/lib/elements/dom-if.js';
 import { html, PolymerElement } from '../node_modules/@polymer/polymer/polymer-element.js';
 import '../views/wpt-404.js';
-import '../views/wpt-results.js';
 import '../views/wpt-interop.js';
+import '../views/wpt-results.js';
 
-class WPTApp extends WPTFlags(TestRunsUIQuery(PolymerElement)) {
+class WPTApp extends PathInfo(WPTFlags(TestRunsUIQuery(PolymerElement))) {
   static get is() { return 'wpt-app'; }
 
   static get template() {
@@ -158,18 +159,6 @@ class WPTApp extends WPTFlags(TestRunsUIQuery(PolymerElement)) {
         type: String,
         computed: '_computePath(subroute.path)',
       },
-      encodedPath: {
-        type: String,
-        computed: 'encodeTestPath(path)'
-      },
-      pathIsATestFile: {
-        type: Boolean,
-        computed: 'computePathIsATestFile(path)'
-      },
-      pathIsASubfolder: {
-        type: Boolean,
-        computed: 'computePathIsASubfolder(path)'
-      },
       structuredSearch: Object,
       interopLoading: Boolean,
       resultsLoading: Boolean,
@@ -246,23 +235,6 @@ class WPTApp extends WPTFlags(TestRunsUIQuery(PolymerElement)) {
 
   _computePath(subroutePath) {
     return subroutePath || '/';
-  }
-
-  encodeTestPath(path) {
-    path = path || '/';
-    console.assert(path.startsWith('/'));
-    let parts = path.split('/').slice(1);
-    parts.push(encodeURIComponent(parts.pop()));
-    return '/' + parts.join('/');
-  }
-
-  computePathIsATestFile(path) {
-    return /(\.(html|htm|py|svg|xhtml|xht|xml)(\?.*)?$)/.test(path);
-  }
-
-  computePathIsASubfolder(path) {
-    return !this.computePathIsATestFile(path)
-      && path && path.split('/').filter(p => p).length > 0;
   }
 
   splitPathIntoLinkedParts(inputPath) {
