+++ webapp/views/wpt-results.js
@@ -4,23 +4,9 @@
  * found in the LICENSE file.
  */
 
-import '../node_modules/@polymer/iron-collapse/iron-collapse.js';
-import '../node_modules/@polymer/iron-icon/iron-icon.js';
-import '../node_modules/@polymer/iron-icons/editor-icons.js';
-import '../node_modules/@polymer/iron-icons/image-icons.js';
-import '../node_modules/@polymer/paper-button/paper-button.js';
-import '../node_modules/@polymer/paper-icon-button/paper-icon-button.js';
-import '../node_modules/@polymer/paper-spinner/paper-spinner-lite.js';
-import '../node_modules/@polymer/paper-styles/color.js';
-import '../node_modules/@polymer/paper-toast/paper-toast.js';
-import '../node_modules/@polymer/paper-tabs/paper-tabs.js';
-import '../node_modules/@polymer/polymer/lib/elements/dom-if.js';
-import '../node_modules/@polymer/polymer/lib/elements/dom-repeat.js';
-import '../node_modules/@polymer/polymer/polymer-element.js';
-import { html } from '../node_modules/@polymer/polymer/polymer-element.js';
 import '../components/info-banner.js';
 import { LoadingState } from '../components/loading-state.js';
-import '../components/path-part.js';
+import '../components/path.js';
 import '../components/test-file-results-table-terse.js';
 import '../components/test-file-results-table-verbose.js';
 import '../components/test-file-results.js';
@@ -34,10 +20,25 @@ import { WPTColors } from '../components/wpt-colors.js';
 import { WPTFlags } from '../components/wpt-flags.js';
 import '../components/wpt-permalinks.js';
 import '../components/wpt-prs.js';
+import '../node_modules/@polymer/iron-collapse/iron-collapse.js';
+import '../node_modules/@polymer/iron-icon/iron-icon.js';
+import '../node_modules/@polymer/iron-icons/editor-icons.js';
+import '../node_modules/@polymer/iron-icons/image-icons.js';
+import '../node_modules/@polymer/paper-button/paper-button.js';
+import '../node_modules/@polymer/paper-icon-button/paper-icon-button.js';
+import '../node_modules/@polymer/paper-spinner/paper-spinner-lite.js';
+import '../node_modules/@polymer/paper-styles/color.js';
+import '../node_modules/@polymer/paper-tabs/paper-tabs.js';
+import '../node_modules/@polymer/paper-toast/paper-toast.js';
+import '../node_modules/@polymer/polymer/lib/elements/dom-if.js';
+import '../node_modules/@polymer/polymer/lib/elements/dom-repeat.js';
+import '../node_modules/@polymer/polymer/polymer-element.js';
+import { html } from '../node_modules/@polymer/polymer/polymer-element.js';
+import { PathInfo } from '../components/path.js';
 
 const TEST_TYPES = ['manual', 'reftest', 'testharness', 'visual', 'wdspec'];
 
-class WPTResults extends WPTColors(WPTFlags(LoadingState(TestRunsUIBase))) {
+class WPTResults extends WPTColors(WPTFlags(PathInfo(LoadingState(TestRunsUIBase)))) {
   static get template() {
     return html`
     <style include="wpt-colors">
